Разделение на модули
Разделен на модули существующий код, добавлены пустые модули, которые потребуются в дальнейшем.
