+++ js/util.js
@@ -1,0 +1,20 @@
+const getRandomNumber = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return (max > min && min >= 0)
+    ? Math.floor(Math.random() * (max - min + 1)) + min
+    : undefined;
+};
+
+const getRandomFloat = (min, max, decimalPoints = 1) => {
+  if (max > min && min >= 0) {
+    const randomNumber = Math.random() * (max - min + 1) + min;
+    return randomNumber.toFixed(decimalPoints);
+  }
+  return undefined;
+};
+
+const getRandomArrayElement = (elements) =>
+  elements[getRandomNumber(0, elements.length - 1)];
+
+export {getRandomNumber, getRandomFloat, getRandomArrayElement};
